+++ src/LRMS.Infrastructure/ReservationManagerApi/Books/Dto/BookReservationModel.cs
@@ -1,3 +1,3 @@
 ﻿namespace LRMS.Infrastructure.ReservationManagerApi.Books.Dto;
 
-public record struct BookReservationModel(int id, int book_id, string name, string phone, string reserved_at);
+public record struct BookReservationModel(long id, long book_id, string name, string phone, string reserved_at);
