+++ src/LRMS.Infrastructure/ReservationManagerApi/Books/Dto/CreateBookReservationDto.cs
@@ -1,3 +1,3 @@
 ﻿namespace LRMS.Infrastructure.ReservationManagerApi.Books.Dto;
 
-public record struct CreateBookReservationDto(int book_id, string name, string phone, string reserved_at);
+public record struct CreateBookReservationDto(long book_id, string name, string phone, string reserved_at);
