+++ src/LRMS.Infrastructure/ReservationManagerApi/Books/IBookReservationRepository.cs
@@ -5,7 +5,7 @@ namespace LRMS.Infrastructure.ReservationManagerApi.Books;
 public interface IBookReservationRepository
 {
     Task CreateBookReservation(CreateBookReservationCommand bookReservationDto, CancellationToken ct = default);
-    Task<GetBookReservationsResponse> GetBookReservations(int? bookId, DateTime? date, int? pageNumber, int? pageSize,
+    Task<GetBookReservationsResponse> GetBookReservations(long? bookId, DateTime? date, int? pageNumber, int? pageSize,
         CancellationToken ct = default);
-    Task DeleteBookReservation(int id, CancellationToken ct = default);
+    Task DeleteBookReservation(long id, CancellationToken ct = default);
 }
