+++ src/LRMS.Infrastructure/ReservationManagerApi/Tables/Dto/CreateTableReservationDto.cs
@@ -1,3 +1,3 @@
 ﻿namespace LRMS.Infrastructure.ReservationManagerApi.Tables.Dto;
 
-public record struct CreateTableReservationDto(int table_id, string name, string phone, string start_at, string end_at);
+public record struct CreateTableReservationDto(long table_id, string name, string phone, string start_at, string end_at);
