+++ src/LRMS.Infrastructure/ReservationManagerApi/Tables/ITableApi.cs
@@ -10,12 +10,12 @@ public interface ITableApi
     Task<HttpResponseMessage> CreateTableReservation([Body] CreateTableReservationDto dto, CancellationToken ct = default);
 
     [Get("/api/v1/tables")]
-    Task<HttpResponseMessage> GetTableReservations([Refit.Query] int? tableId, [Refit.Query] string? activeAt,
+    Task<HttpResponseMessage> GetTableReservations([Refit.Query] long? tableId, [Refit.Query] string? activeAt,
         [Refit.Query] int? pageNumber, [Refit.Query] int? pageSize, CancellationToken ct = default);
 
     [Get("/api/v1/tables/{tableId}/slots")]
-    Task<HttpResponseMessage> GetTableReservationSlots(int tableId, [Refit.Query] string date, CancellationToken ct = default);
+    Task<HttpResponseMessage> GetTableReservationSlots(long tableId, [Refit.Query] string date, CancellationToken ct = default);
 
     [Delete("/api/v1/tables/{id}")]
-    Task<HttpResponseMessage> DeleteTableReservation(int id, CancellationToken ct = default);
+    Task<HttpResponseMessage> DeleteTableReservation(long id, CancellationToken ct = default);
 }
